re-write code to a separates files, fix errors in eslint.js
