+++ src/css/start.css
@@ -114,3 +114,8 @@
   text-align: center;
   align-items: center;
 }
+#map {
+  width: 500px;
+  height: 400px;
+  margin-left: 500px;
+}
